feat: support method

+++ cmd/reflect-shape-dump/main.go
@@ -1,12 +1,14 @@
 package main
 
 import (
+	"bytes"
 	"context"
 	"io/ioutil"
 	"log"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strconv"
 	"strings"
 	"text/template"
 )
@@ -36,11 +38,36 @@ func run() error {
 	e.RevisitArglist = true
 	e.ArglistLookup = arglist.NewLookup()
 
+{{ if eq .Kind "Func" }}
+	s := e.Extract({{.Qualifier}}.{{.Name}})
+	return reflectshape.Fdump(os.Stdout, s)
+{{ else if eq .Kind "Method" }}
+	ob := &{{.Qualifier}}.{{.Recv}}{}
+	s := e.Extract(ob.{{.Name}})
+	return reflectshape.Fdump(os.Stdout, s)
+{{ else }}
 	var target func() {{.Qualifier}}.{{.Name}}
 	s := e.Extract(target).(reflectshape.Function)
 	return reflectshape.Fdump(os.Stdout, s.Returns.Values[0])
+{{ end }}
 }
 `
+
+var keep bool
+
+type Kind string
+
+const (
+	KindOther  Kind = ""
+	KindMethod Kind = "Method"
+	KindFunc   Kind = "Func"
+)
+
+func init() {
+	if v, err := strconv.ParseBool(os.Getenv("KEEP")); err == nil {
+		keep = v
+	}
+}
 
 func main() {
 	log.SetFlags(0)
@@ -54,15 +81,42 @@ func run() error {
 		log.Fatalf("please run <cmd> <fullpath>\n\t e.g. go run <> github.com/podhmo/reflect-shape.Struct\n")
 	}
 	t := template.Must(template.New("code").Parse(code))
+
 	fullpath := os.Args[1]
 	parts := strings.Split(fullpath, ".")
 	path := strings.Join(parts[:len(parts)-1], ".")
 	name := parts[len(parts)-1]
+	if strings.LastIndex(path, "/") < strings.LastIndex(path, ".") {
+		path = strings.Join(parts[:len(parts)-2], ".")
+		name = strings.Join(parts[len(parts)-2:], ".")
+	}
 
 	log.Printf("check by go list.\tpath=%q\tname=%q", path, name)
 	ctx := context.Background()
-	if err := exec.CommandContext(ctx, "go", "list", path).Run(); err != nil {
-		return err
+	{
+		cmd := exec.CommandContext(ctx, "go", "list", path)
+		if err := cmd.Run(); err != nil {
+			return err
+		}
+	}
+
+	log.Println("check by go doc.")
+	buf := new(bytes.Buffer)
+	{
+		cmd := exec.CommandContext(ctx, "go", "doc", "-short", fullpath)
+		cmd.Stdout = buf
+		if err := cmd.Run(); err != nil {
+			return err
+		}
+	}
+
+	kind := KindOther
+	if strings.HasPrefix(strings.TrimSpace(buf.String()), "func") {
+		if strings.HasPrefix(strings.TrimSpace(buf.String()), "func (") {
+			kind = KindMethod
+		} else {
+			kind = KindFunc
+		}
 	}
 
 	d, err := ioutil.TempDir(".", ".reflect-shape")
@@ -70,6 +124,10 @@ func run() error {
 		return err
 	}
 	defer func() {
+		if keep {
+			return
+		}
+
 		log.Println("remove all", d)
 		if err := os.RemoveAll(d); err != nil {
 			log.Println("!? something wrong in os.RemoveAll(),", err)
@@ -83,10 +141,18 @@ func run() error {
 	defer f.Close()
 
 	log.Println("create", filepath.Join(d, "main.go"))
-	if err := t.Execute(f, map[string]string{
+	recvAndName := strings.SplitN(name, ".", 2)
+	recv := ""
+	if len(recvAndName) > 1 {
+		recv = recvAndName[0]
+		name = recvAndName[1]
+	}
+	if err := t.Execute(f, map[string]interface{}{
 		"Path":      path,
 		"Qualifier": "x",
+		"Recv":      recv,
 		"Name":      name,
+		"Kind":      kind,
 	}); err != nil {
 		return err
 	}
